+++ src/main/scala/com/scalamandra/service/Service.scala
@@ -1,11 +1,12 @@
 package com.scalamandra.service
 
 import com.scalamandra.model.HttpException
+import com.scalamandra.utils.HttpExceptionUtils
 
 import scala.concurrent.Future
 
 trait Service {
 
-  protected final def error[T <: HttpException](obj: T): Future[Left[T, Nothing]] = HttpException.error(obj)
+  protected final def error[T <: HttpException](obj: T): Future[Left[T, Nothing]] = HttpExceptionUtils.error(obj)
 
 }
